requirements.txt
Flask
gunicorn
torch
torchvision
pandas
Pillow
numpy
gdown
requests
